v1.4

+++ log_process.go
@@ -144,8 +144,8 @@ func MemStat() MemStatus {
 	sysInfo := new(syscall.Sysinfo_t)
 	err := syscall.Sysinfo(sysInfo)
 	if err == nil {
-		mem.All = sysInfo.Totalram * uint32(syscall.Getpagesize())
-		mem.Free = sysInfo.Freeram * uint32(syscall.Getpagesize())
+		mem.All = sysInfo.Totalram * uint64(syscall.Getpagesize())
+		mem.Free = sysInfo.Freeram * uint64(syscall.Getpagesize())
 		mem.Used = mem.All - mem.Free
 	}
 	return mem
